requirements.txt
azure-cognitiveservices-vision-computervision
msrest
openai
deepface
tensorflow
keras
flask
flask_sqlalchemy
flask_wtf
flask_login
numpy
scikit-learn
python-dotenv
